SpringMvcClient support discovery config registration
